requirements.txt
python==3.10.0
django==3.1
django-phone-field==1.8.1